refactor: Move email validation into its own file
